1.增加记录骑行的点显示在地图连成一条折线，并且地图按一定的缩放比列把这条线全部显示在地图上。2.补全仪表盘的ui

+++ app/src/main/java/com/app/iriding/util/SqliteUtil.java
@@ -29,8 +29,8 @@ public class SqliteUtil {
         return cyclingRecords;
     }
 
-    //得到某一次的骑行记录
-    public CyclingRecord selectSingleCyclingRecord(int id){
+    // 得到某一次的骑行记录
+    public CyclingRecord selectSingleCyclingRecord(long id){
         CyclingRecord cyclingRecord = new Select().from(CyclingRecord.class)
                 .where("Id = ?", id)
                 .executeSingle();
